Move octous file loading and exporting to separate helper files
