+++ ode-logic/ode-logic.h
@@ -15,5 +15,6 @@
 #include "ode/design-management/Component.h"
 #include "ode/design-management/Design.h"
 #include "ode/design-management/OctopusFile.h"
+#include "ode/design-management/octopus-api-helpers.h"
 #include "ode/animation/DocumentAnimation.h"
 #include "ode/animation/animate.h"
