+++ tools/design-editor/ode/widgets/DesignEditorControlsWidget.cpp
@@ -149,96 +149,7 @@ void drawControlsWidget(DesignEditorDesign &design,
 
             if (ImGuiFileDialog::Instance()->IsOk()) {
                 const std::string filePathName = ImGuiFileDialog::Instance()->GetFilePathName();
-                const std::string fileName = (FilePath(filePathName)).filename();
-
-                const size_t extPos = fileName.rfind('.', fileName.length());
-                const std::string cleanFileName = (extPos == std::string::npos) ? fileName : fileName.substr(0, extPos);
-
-                OctopusFile octopusFile;
-                octopusFile.add("Octopus", " is universal design format. opendesign.dev.", MemoryFileSystem::CompressionMethod::NONE);
-
-                octopus::OctopusManifest manifest;
-                manifest.version = OCTOPUS_MANIFEST_VERSION;
-                manifest.origin.name = "OD Internal Design Editor";
-                manifest.origin.version = "1.0.0.";
-                manifest.name = cleanFileName;
-
-                for (const DesignEditorComponent &component : design.components) {
-                    const ODE_StringRef &componentId = component.id;
-
-                    ODE_String octopusString;
-                    if (ode_component_getOctopus(component.component, &octopusString) == ODE_RESULT_OK) {
-                        const std::string componentFileName = std::string("octopus-") + ode_stringDeref(componentId) + ".json";
-                        if (!octopusFile.add(componentFileName, std::string(octopusString.data, octopusString.length), MemoryFileSystem::CompressionMethod::DEFLATE).has_value()) {
-                            continue;
-                        }
-
-                        octopus::Octopus octopus;
-                        octopus::Parser::parse(octopus, octopusString.data);
-
-                        octopus::Component &octopusComponent = manifest.components.emplace_back();
-                        octopusComponent.id = octopus.id;
-                        octopusComponent.name = octopus.content->name;
-                        octopusComponent.status = octopus::Status {
-                            octopus::Status::Value::READY,
-                            nonstd::nullopt,
-                            0.0 };
-                        octopusComponent.bounds = octopus::Bounds {
-                            component.metadata.position.x,
-                            component.metadata.position.y,
-                            octopus.dimensions->width,
-                            octopus.dimensions->height };
-                        octopusComponent.location = octopus::ResourceLocation {
-                            octopus::ResourceLocation::Type::RELATIVE,
-                            componentFileName,
-                            nonstd::nullopt,
-                            nonstd::nullopt };
-                        octopusComponent.assets = octopus.content.has_value() ? listAllAssets(*octopus.content) : octopus::Assets{};
-
-                        for (const octopus::AssetImage &assetImage : octopusComponent.assets->images) {
-                            if (assetImage.location.path.has_value()) {
-                                ODE_MemoryBuffer imageData;
-                                if (ode_pr1_design_exportPngImage(design.imageBase, ode_stringRef(assetImage.location.path.value()), &imageData) == ODE_RESULT_OK) {
-                                    if (!octopusFile.add(assetImage.location.path.value(), std::string(static_cast<const char *>(imageData.data), imageData.length), MemoryFileSystem::CompressionMethod::NONE)) {
-                                        // TODO: Error saving PNG to octopus file
-                                        continue;
-                                    }
-                                }
-                            }
-                        }
-                        for (octopus::AssetFont &assetFont : octopusComponent.assets->fonts) {
-                            ODE_MemoryBuffer fontData;
-                            if (ode_pr1_design_exportFontBytes(design.design, ODE_StringRef {assetFont.name.data(), (int)assetFont.name.size()}, &fontData) == ODE_RESULT_OK) {
-                                if (fontData.data) {
-                                    const std::string fileExtension = fontFileExtension(fontData);
-                                    if (!fileExtension.empty()) {
-                                        assetFont.location = octopus::ResourceLocation {
-                                            octopus::ResourceLocation::Type::RELATIVE,
-                                            "fonts/"+assetFont.name+"."+fileExtension,
-                                            nonstd::nullopt,
-                                            nonstd::nullopt };
-                                        if (!octopusFile.exists(assetFont.location->path.value())) {
-                                            if (!octopusFile.add(assetFont.location->path.value(), std::string(static_cast<const char *>(fontData.data), fontData.length), MemoryFileSystem::CompressionMethod::NONE)) {
-                                                // TODO: Error saving Font to octopus file
-                                                continue;
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                std::string manifestJson;
-                if (octopus::ManifestSerializer::serialize(manifestJson, manifest) == octopus::ManifestSerializer::Error::OK) {
-                    if (octopusFile.add("octopus-manifest.json", manifestJson, MemoryFileSystem::CompressionMethod::DEFLATE).has_value()) {
-                        const bool isSaved = octopusFile.save(filePathName);
-                        if (!isSaved) {
-                            fprintf(stderr, "Internal error (saving Octopus json to filesystem)\n");
-                        }
-                    }
-                }
+                ode_pr1_saveDesignToFile_Media(design.design, ode_stringRef(filePathName), design.imageBase);
             }
 
             ImGuiFileDialog::Instance()->Close();
